+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.4.0 - 15/Jan/2018
 
+* BeanSpec - add parent() method
 * BeanSpec - add method to get bean spec of all fields #26
 * Support dynamic named provider #25
 * Do not cache provider for simple types without inject tag #24